Added a morning-routine script
Since I didn't regulary read the mails from IFTTT i decided
to create a script that runs on startup and is difficult to
miss. The script launches the tools in a subshell and forces
me to do the routine in the specified order.

Since task, mutt and anki are now started from the script,
they are no longer required in the .xinitrc.