Make ssh work
This fix, will make the ssh login work on the shell. The keyring gets
unlocked on startup and everything is fine for ssh logins.

This does not fix the problems with owncloud, but I believe they are an
issue with owncloud or it's dependencies (See [2][2] and [3][3]).

It is a big issue and I spent the last 2 days investigating it. It
seems that owncloud has no connection to the keychain via dbus and one
solution is to not start any gnome-keyring-deamon manually. (See [1][1])
But then ssh would stop working again.

So I'm basically waiting for owncloud to release a new version/updated
dependencies. Until then I will have to enter the owncloud-client
password by myself everytime I start it.

 * [1]: https://faq.i3wm.org/question/5168/owncloud-applet-asks-for-password-when-starting-up.1.html
 * [2]: https://github.com/owncloud/client/issues/3726
 * [3]: https://github.com/owncloud/client/issues/1590

Issue-Link: https://gitlab.com/phansch/dotfiles/issues/24