[script] Add focus_or_launch_browser script
This script will launch firefox and focus the window where it has been
launched. I want to try and see if I can open Firefox only if I really
need it. Like a tool that I grab only when I need it and not otherwise.

Firefox is the biggest source of distractions.