Add startup script to replace xinitrc
This script is executed from i3.
The stuff was originally in xinitrc because I thought it would be nice
to have it independent from the window manager. However, I've been using
i3 for a couple years now and have no plans of replacing it.

Additionally, one could argue that a shell script is even more
independent from the window manager that xinitrc.