vim: Add auto-pairs plugin
I'm still conflicted on this, because sometimes it's not what I want,
but sometimes it makes editing a lot quicker.

I guess this is because the current keyboard I have in my home office
has the default locations for `{[()]}`, which are hard to type on this
particular keyboard.

On my Ergodox at work, these characters are _much_ easier to type and I
probably wouldn't need `auto-pairs` on the Ergodox.